decomp: fix 9 confirmed fidelity bugs from bug-hunt wave 2
Found via objdiff report-analyze (OFFSET_SWAP / CONTROL_FLOW candidates),
each hand-decomped and coordinator-re-scored:

- HamDriver LayerClip::Replace: `delete this` + return-value control flow
  (decomp deleted the member clip via mClip.Ptr() and always returned true;
  retail deletes the object itself) -> memory-management fidelity bug. -> 100
- CharClipSet::ListDrawChildren: wrong arg to the child draw call. 71.4 -> 100
- Automator::OnMsg (UI.cpp): wrong argument. 99.98 -> 100
- op (BlockMgr ReadError): emitted a literal placeholder instead of the
  function call. 96.5 -> 100
- MoveMgr::SaveRoutineVariants: bind the ternary result by const-reference
  instead of by value. 95.3 -> 100
- RndText::AcquireFontMap: wrong virtual accessor + wrong comparand in the
  font-map cache-lookup predicate. 85.4 -> 95.3
- RndText::UpdateScrollOffsets: missing mScrollPos reset. 99.2 -> 99.97
- SpeechMgr::AddDynamicRuleWord: restore the second wchar_t[1024] stack buffer
  (decomp aliased one buffer for two uses). 97.6 -> 98.4
- WavReader::WavReader: correct a signedness conversion. 96.7 -> 99.4

Residuals are register-allocation floors, not divergence.

+++ src/system/synth/WavReader.cpp
@@ -17,7 +17,7 @@ WavReader::WavReader(File *file, StandardStream *stream) {
     for (int i = 0; i < mInWaveFile->NumMarkers(); i++) {
         WaveFileMarker &wfm = mInWaveFile->Markers()[i];
         int frame = wfm.mFrame;
-        float posMS = (float)frame * 1000.0f / (float)mInWaveFile->mSamplesPerSec;
+        float posMS = (float)frame * 1000.0f / (float)(int)mInWaveFile->mSamplesPerSec;
         Marker marker(wfm.mName);
         marker.position = frame;
         marker.posMS = posMS;
